part 4 osa 4.21

+++ part4/controllers/blogs.js
@@ -21,13 +21,27 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-
-
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (!blogToDelete) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (decodedToken.id !== blogToDelete.user.toString()) {
+    return response.status(403).json({ error: 'forbidden: cannot delete other user\'s blog' })
+  }
 
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
+
 
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
